// backend/src/libs/mikro-orm/crud/enums/find-operator.ts
export enum FindOperator {
  EQ = 'eq',
  NE = 'ne',
  GT = 'gt',
  LT = 'lt',
  GTE = 'gte',
  LTE = 'lte',
  CONTAINS = 'contains',
  IN = 'in',
  NOT_IN = 'notIn',
  IS_TRUE = 'isTrue',
  IS_FALSE = 'isFalse',
  IS_NULL = 'isNull',
  IS_NOT_NULL = 'isNotNull',
  SOME = 'some',
  EVERY = 'every',
  NONE = 'none',
  LIKE = 'like',
  ILIKE = 'ilike',
}